requirements.txt
fastapi
uvicorn[standard]
python-jose
passlib
python-multipart
sqlalchemy
pymysql


